11.02.2022 - rev2: замечания rev1 исправлены

+++ src/main/java/ru/practicum/shareit/item/service/ItemServiceImpl.java
@@ -4,6 +4,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
+import ru.practicum.shareit.exception.ResourceNotFoundException;
+import ru.practicum.shareit.exception.ForbiddenResourceException;
 import ru.practicum.shareit.item.dto.ItemDto;
 import ru.practicum.shareit.item.dto.MapperItem;
 import ru.practicum.shareit.item.model.Item;
@@ -13,6 +15,7 @@ import ru.practicum.shareit.user.dto.UserDto;
 import ru.practicum.shareit.user.model.User;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -37,13 +40,30 @@ public class ItemServiceImpl implements ItemService {
     public ItemDto updateItem(Long itemId, ItemDto itemDto, Long userId) {
         log.info("Получен запрос на обновление вещи с ID " + itemId + " от пользователя с ID " + userId);
         Item item = MapperItem.toItem(itemDto);
-        return MapperItem.toItemDto(itemRepository.updateItem(itemId, item, userId));
+        List<Item> itemsRepository = itemRepository.getItemsByUser(userId);
+        if (itemsRepository.stream().anyMatch(i -> i.getId().equals(itemId))) {
+            if (itemsRepository.stream().filter(i -> i.getId().equals(itemId))
+                    .anyMatch(i -> i.getOwner().getId().equals(userId))) {
+                return MapperItem.toItemDto(itemRepository.updateItem(itemId, item, userId));
+            } else {
+                throw new ForbiddenResourceException("Пользователь ID " + userId + " пытался обновить вещь "
+                        + itemId + " данная вещь ему не принадлежит");
+            }
+        } else {
+            throw new ResourceNotFoundException("Вещь с ID " + itemId + " не найдена");
+        }
     }
 
     @Override
     public ItemDto getItemById(Long itemId, Long userId) {
         log.info("Получен запрос на поиск вещи с ID " + itemId + " от пользователя с ID " + userId);
-        return MapperItem.toItemDto(itemRepository.getItemById(itemId, userId));
+        Optional<Item> itemOptional = itemRepository.getItemById(itemId, userId);
+        if (itemOptional.isPresent()) {
+            return MapperItem.toItemDto(itemOptional.get());
+        } else {
+            throw new ResourceNotFoundException("Вещь с ID " + itemId + " не найдена, по запросу " +
+                    "пользователя с ID " + userId);
+        }
     }
 
     @Override
@@ -61,12 +81,17 @@ public class ItemServiceImpl implements ItemService {
     @Override
     public void deleteItemById(Long itemId, Long userId) {
         log.info("Получен запрос на удаление вещи с ID " + itemId + " от пользователя с ID " + userId);
-        itemRepository.deleteItemById(itemId, userId);
+        List<Item> itemsUser = itemRepository.getItemsByUser(userId);
+        if (itemsUser.stream().anyMatch(i -> i.getId().equals(itemId))) {
+            itemRepository.deleteItemById(itemId, userId);
+        } else {
+            throw new ResourceNotFoundException("Вещь с ID " + itemId + " не найдена");
+        }
     }
 
     @Override
     public List<ItemDto> getAllItems() {
-        log.info("Получен запрос на получение списока всех вещей");
+        log.info("Получен запрос на получение списка всех вещей");
         return itemRepository.getAllItems().stream().map(MapperItem::toItemDto).collect(Collectors.toList());
     }
 }
